vectorization

+++ Step0/neuron.cpp
@@ -17,7 +17,7 @@ float evalNeuron(
 )
 {
 	float result = 0;
-	for(int i = 0; i < inputSize; i++)
+	for(size_t i = 0; i < inputSize; i++)
 		result += input[i] * weights[i * neuronCount + neuronId];
 	result += bias;
